feat: implement trigger-form component
