+++ dashboard/src/components/input/device-and-pin-input/device-and-pin-input.tsx
@@ -5,8 +5,6 @@ import AutocompletedSearch from '@/components/autocompleted-search/autocompleted
 const DeviceAndPinInputs = ({
     allDevices,
     initialDeviceName = '',
-    deviceIdLable,
-    deviceIdValue,
     deviceNameLabel,
     pinLabel,
     pinValue,
@@ -16,8 +14,6 @@ const DeviceAndPinInputs = ({
 }: {
     allDevices: DeviceItem[];
     initialDeviceName?: string;
-    deviceIdLable: string;
-    deviceIdValue: number;
     deviceNameLabel: string;
     pinLabel: string;
     pinValue: string;
@@ -36,26 +32,19 @@ const DeviceAndPinInputs = ({
     }, [currentDeviceId]);
 
     return (
-        <>
-            <div className="form-group mt-3">
-                <label>{deviceIdLable}</label>
-                <input
-                    className="form-control"
-                    disabled
-                    value={deviceIdValue}
-                />
-            </div>
-            <div className="form-group mt-3">
-                <label>{deviceNameLabel}</label>
+        <div className="d-flex f-wrap w-100 mb-3">
+            <div className="form-group w-50 pe-3">
+                <label className="mb-1">{deviceNameLabel}</label>
                 <AutocompletedSearch
                     currentValue={deviceName}
                     updateCurrentValue={(newValue) => setDeviceName(newValue)}
                     allSuggestions={allDevices.map(({ name }) => name)}
                 />
             </div>
-            <div className="form-group mt-3">
-                <label>{pinLabel}</label>
+            <div className="form-group w-50 ps-3">
+                <label className="mb-1">{pinLabel}</label>
                 <select
+                    className="form-select"
                     value={pinValue}
                     onChange={(e) => {
                         const newSourcePin = e.target.value;
@@ -81,7 +70,7 @@ const DeviceAndPinInputs = ({
                     )}
                 </select>
             </div>
-        </>
+        </div>
     );
 };
