+++ dashboard/src/pages/trigger/trigger-form/trigger-form.scss
@@ -1,0 +1,3 @@
+.trigger-form-wrapper {
+  width: 720px;
+}
